refactor: dto and entities
fix: typo

refactor: prisma module

refactor: transfer to service

feat: import

refactor: naming convention

refactor: schema

refactor: folder structure
