+++ src/todos/dto/create-todo.input.ts
@@ -1,7 +1,11 @@
 import { InputType, Int, Field } from '@nestjs/graphql';
+import { Prisma } from '@prisma/client';
 
 @InputType()
-export class CreateTodoInput {
-  @Field(() => Int, { description: 'Example field (placeholder)' })
-  exampleField: number;
+export class CreateTodoInput implements Prisma.TodoCreateInput {
+  @Field(() => String)
+  title: string;
+
+  @Field(() => String, { nullable: true })
+  description?: string;
 }
