+++ src/todos/dto/update-todo.input.ts
@@ -5,4 +5,10 @@ import { InputType, Field, Int, PartialType } from '@nestjs/graphql';
 export class UpdateTodoInput extends PartialType(CreateTodoInput) {
   @Field(() => Int)
   id: number;
+
+  @Field(() => String, { nullable: true })
+  title?: string;
+
+  @Field(() => String, { nullable: true })
+  description?: string;
 }
