+++ src/todos/entities/todo.entity.ts
@@ -1,7 +1,22 @@
 import { ObjectType, Field, Int } from '@nestjs/graphql';
-
+import { Todo as TodoDB } from '@prisma/client';
 @ObjectType()
-export class Todo {
-  @Field(() => Int, { description: 'Example field (placeholder)' })
-  exampleField: number;
+export class Todo implements TodoDB {
+  @Field(() => Int)
+  id: number;
+
+  @Field(() => String)
+  title: string;
+
+  @Field(() => String, { nullable: true })
+  description?: string | null;
+
+  @Field(() => Boolean)
+  isCompleted: boolean;
+
+  @Field(() => Date)
+  createdAt: Date;
+
+  @Field(() => Date)
+  updatedAt: Date;
 }
