+++ src/todos/todos.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TodosService } from './todos.service';
 import { TodosResolver } from './todos.resolver';
+import { PrismaModule } from 'src/prisma/prisma.module';
 
 @Module({
+  imports: [PrismaModule],
   providers: [TodosResolver, TodosService],
 })
 export class TodosModule {}
