+++ src/todos/todos.service.ts
@@ -1,26 +1,30 @@
 import { Injectable } from '@nestjs/common';
-import { CreateTodoInput } from './dto/create-todo.input';
-import { UpdateTodoInput } from './dto/update-todo.input';
+
+import { PrismaService } from '../prisma/prisma.service';
+
+import { Prisma } from '@prisma/client';
 
 @Injectable()
 export class TodosService {
-  create(createTodoInput: CreateTodoInput) {
-    return 'This action adds a new todo';
+  constructor(private prisma: PrismaService) {}
+
+  async create(data: Prisma.TodoCreateInput) {
+    return await this.prisma.todo.create({ data });
   }
 
-  findAll() {
-    return `This action returns all todos`;
+  async findAll() {
+    return await this.prisma.todo.findMany();
   }
 
-  findOne(id: number) {
-    return `This action returns a #${id} todo`;
+  async findOne(id: number) {
+    return await this.prisma.todo.findUnique({ where: { id } });
   }
 
-  update(id: number, updateTodoInput: UpdateTodoInput) {
-    return `This action updates a #${id} todo`;
+  async update(id: number, data: Prisma.TodoUpdateInput) {
+    return await this.prisma.todo.update({ where: { id }, data });
   }
 
-  remove(id: number) {
-    return `This action removes a #${id} todo`;
+  async remove(id: number) {
+    return await this.prisma.todo.delete({ where: { id } });
   }
 }
